Sync issue
Sync currently has the issue, that if the Folder doesn't contain any images it exits the shell.
This leads to the app status not being checkt and new config files not beeing loaded.
Removing this one exit doesn't seem to cause errors, even if the folder is empty.

There are 2 more exit statements in the file that should be thought about. Maybe replace them with if statements?