Added actual restart
Currently if the config is changed and pi-timolo is running there is no real restart. 
I added a stop command before the start command to make it a real restart.
This is necessary, because python creates a .pyc file for the config and even if config.py is changed no changes take place until a actual restart has happened.