Added a Wipe function to sync.sh
I noticed syncing becoming really slow at about 400 files. 
So I added a modified remote-config-function to wipe the sync dir if a certain file  exists in the Google drive.
Since the files are already synced when the function is called there should be no dataloss.

This function can be quite handy if your pi is miles away and you don't want the sd card to fill up ;) and it speeds up sync.